% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_module.R
\name{imp_module}
\alias{imp_module}
\title{imp_module() function}
\usage{
imp_module(
  smp_name = smp_name,
  meta = meta,
  species_upd = species_upd,
  pc = pc
)
}
\arguments{
\item{X}{after preprocess step}
}
\description{
This function is imputation module for species detected only from WGS
}
\examples{
imp_module())


}
\keyword{imputation}
\keyword{module}
