% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microimpute.R
\name{microimpute}
\alias{microimpute}
\title{microimpute()}
\usage{
microimpute(
  X = smpdata,
  meta = meta,
  preprocess = FALSE,
  normalized = TRUE,
  sep = ","
)
}
\arguments{
\item{X}{abundance}

\item{meta}{meta data}

\item{preprocess}{preprocessed or not}

\item{normalized}{normalized or not}

\item{sep}{separator}
}
\value{
A imputed matrix of dimension n_taxa x n_samples
}
\description{
This function is the main function of MicroImpute R package
}
\examples{
microimpute()


}
\keyword{MicroImpute}
\keyword{main}
\keyword{module}
