% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{rownameset}
\alias{rownameset}
\title{preprocess() function}
\usage{
rownameset(df)
}
\arguments{
\item{abd_loc}{: abundance data location for import}

\item{meta}{: metadata location for import}

\item{sep}{: separator for data}
}
\description{
This function preprocess the input data and transform the data for imputation
}
\examples{
preprocess()


}
\keyword{preprocess}
