% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upd_module.R
\name{upd_module}
\alias{upd_module}
\title{upd_module() function}
\usage{
upd_module(X = lmm_input)
}
\arguments{
\item{X}{after preprocess step}
}
\description{
This function is update module for species detected from both platforms (16S and WGS)
}
\examples{
imp_module())


}
\keyword{imputation}
\keyword{module}
